Replace Dexador with Drakma; add AS/UTIL
… which was spurred by an INVALID-VERSION error
from Dexador’s use of fast-http. Drakma seems to
Just Work™, though, so might as well switch.

This is done by abstracting HTTP gets into a util
subpackage, since both AS/JLD & AS need HTTP reqs.

This also defines the FETCH-ERROR, for when
an HTTP error is given during a FETCH call.